Of course I forgot to commit new files.

+++ src/main/java/me/ichun/mods/googlyeyes/common/tracker/GooglyTracker.java
@@ -19,7 +19,7 @@ public class GooglyTracker
     public final float renderChance;
 
     public boolean shouldUpdate = true;
-    public long lastUpdateRequest;
+    public int lastUpdateRequest;
 
     public double motionX;
     public double motionY;
